requirements.txt
asgiref==3.5.2
dj-places==5.1.0
Django==4.1.3
django-crispy-forms==1.14.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
